feat: create ts file for create config script
