+++ packages/untp-test-suite/src/interfaces/cli/index.ts
@@ -1,0 +1,1 @@
+export * from './createConfigFile';
