+++ packages/untp-test-suite/src/interfaces/index.ts
@@ -1,3 +1,3 @@
 // export * from './api'
-// export * from './cli'
+export * from './cli'
 // export * from './service'
